requirements.txt
# Core Audio Processing
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# LLM Integration
ollama>=0.1.9
requests>=2.31.0

# Audio File Handling
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# CLI and Utilities
click>=8.1.7
rich>=13.5.2
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Advanced Speaker Diarization (Optional - requires Python 3.10+)
# Uncomment the next lines if you want advanced diarization:
# pyannote.audio>=3.1.1
# speechbrain>=0.5.0

# Optional dependencies (install separately if needed)
# presidio-analyzer>=2.2.33  # For PHI redaction
# presidio-anonymizer>=2.2.33  # For PHI redaction
# cryptography>=41.0.0  # For security features